move cluster_info::Node to separate module
